PollyServer/requirements.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
Django==5.0.1
django-filter==23.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
Flask==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.4
mysqlclient==2.2.1
pytz==2023.3.post1
sqlparse==0.4.4
Werkzeug==3.0.1
